PR feedback refactor modules
